requirements.txt
# Core simulation dependencies
numpy
networkx==1.11
python-dateutil
Faker

# Visualization dependencies (optional - requires: brew install graphviz)
matplotlib>=3.1.2
pygraphviz
powerlaw

# Neo4j integration dependencies
neo4j>=5.0.0
tqdm